feat(aapd-255): make removeFiles async
